+++ jetcache-test/src/test/java/com/alicp/jetcache/anno/config/ConfigTest.java
@@ -9,7 +9,7 @@ import com.alicp.jetcache.embedded.EmbeddedCacheConfig;
 import com.alicp.jetcache.embedded.LinkedHashMapCacheBuilder;
 import com.alicp.jetcache.external.ExternalCacheConfig;
 import com.alicp.jetcache.support.*;
-import com.alicp.jetcache.test.MockRemoteCacheBuilder;
+import com.alicp.jetcache.external.MockRemoteCacheBuilder;
 import com.alicp.jetcache.test.beans.MyFactoryBean;
 import org.junit.Assert;
 import org.junit.Test;
