+++ jetcache-test/src/test/resources/beans3.0.xml
@@ -35,7 +35,7 @@
     <bean id="localCacheBuilder" class="com.alicp.jetcache.embedded.LinkedHashMapCacheBuilder">
         <property name="keyConvertor" ref="fastjsonKeyConvertor"/>
     </bean>
-    <bean id="remoteCacheBuilder" class="com.alicp.jetcache.test.MockRemoteCacheBuilder">
+    <bean id="remoteCacheBuilder" class="com.alicp.jetcache.external.MockRemoteCacheBuilder">
         <property name="keyPrefix" value="remote1"/>
         <property name="keyConvertor" ref="fastjsonKeyConvertor"/>
         <property name="valueEncoder">
@@ -45,7 +45,7 @@
             <util:constant static-field="com.alicp.jetcache.support.KryoValueDecoder.INSTANCE"/>
         </property>
     </bean>
-    <bean id="remoteCacheBuilder2" class="com.alicp.jetcache.test.MockRemoteCacheBuilder">
+    <bean id="remoteCacheBuilder2" class="com.alicp.jetcache.external.MockRemoteCacheBuilder">
         <property name="keyPrefix" value="remote2"/>
         <property name="keyConvertor" ref="fastjsonKeyConvertor"/>
         <property name="valueEncoder">
